pyproject.toml
[tool.poetry]
name = "rec-travels"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.4"}
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.13.3"
pydantic-settings = "^2.6.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.2.0"
pytz = "^2024.2"
sqladmin = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
